exampleofgoofbackend/POVBackend-prod/requirements.txt
asgiref==3.8.1
boto3==1.24.35
b2sdk==2.4.1
botocore==1.27.35
certifi==2024.6.2
cloudinary==1.41.0
Django==4.2
djangorestframework==3.14.0
idna==3.7
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.4
jmespath==1.0.1
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
pytz==2024.1
s3transfer==0.6.2
six==1.16.0
sqlparse==0.5.0
urllib3==1.26.19
stripe==10.1.0
google-auth==2.31.0
pyjwt==2.8.0
requests==2.32.3
cryptography==42.0.8
firebase-admin==6.5.0
typing_extensions==4.12.2
uritemplate==4.1.1
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
gunicorn==22.0.0
whitenoise==6.7.0
google-analytics-data==0.18.12
google-cloud-bigquery==3.26.0
pyotp==2.9.0
celery==5.4.0
python-dotenv==1.0.1
scikit-learn ==1.6.1
numpy==1.26.4
scikit-surprise==1.1.4
pandas==2.2.3
implicit==0.7.2
tqdm==4.67.1
annoy==1.17.3
scipy
transformers==4.49.0
redis==5.0.2
sentry-sdk==2.22.0